VIH-6173 move hearing controls to separate component
