+++ VideoWeb/VideoWeb/ClientApp/src/app/waiting-space/hearing-controls/hearing-controls.component.html
@@ -1,0 +1,100 @@
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-one-half" id="outgoingFeedDiv">
+    <button
+      id="outgoingFeedButton"
+      type="button"
+      class="govuk-button govuk-button--secondary reduce-margin govuk-!-margin-right-4"
+      data-module="govuk-button"
+      (click)="toggleView()"
+    >
+      {{ selfViewOpen ? 'Close' : 'Open' }} self-view
+    </button>
+
+    <video
+      [ngClass]="{ showVideo: selfViewOpen, hideVideo: !selfViewOpen }"
+      id="outgoingFeedVideo"
+      [muted]="true"
+      [srcObject]="this.outgoingStream"
+      autoplay
+      height="auto"
+    >
+      Your browser does not support Video
+    </video>
+
+    <ng-container *ngIf="!remoteMuted">
+      <span *ngIf="!audioMuted; else unmuteButton">
+        <button
+          id="muteButton"
+          class="govuk-button reduce-margin"
+          data-module="govuk-button"
+          type="button"
+          (click)="toggleMute()"
+          aria-label="Mute Microphone Button"
+        >
+          Mute
+          <span class="govuk-button__start-icon">
+            <img id="mute-microphone-img" class="microphoneIcon" src="/assets/images/mic_on.png" alt="Mute Microphone Image" />
+          </span>
+        </button>
+      </span>
+    </ng-container>
+
+    <ng-template #unmuteButton>
+      <span>
+        <button
+          id="unmuteButton"
+          class="govuk-button govuk-button--warning reduce-margin"
+          data-module="govuk-button"
+          type="button"
+          (click)="toggleMute()"
+          aria-label="Unmute Microphone Button"
+        >
+          Unmute
+
+          <span class="govuk-button__start-icon microphoneIcon">
+            <img id="unmute-microphone-img" class="microphoneIcon" src="/assets/images/mic_off.png" alt="Unmute Microphone Image" />
+          </span>
+        </button>
+      </span>
+    </ng-template>
+    <!-- Remote muted -->
+    <span [ngClass]="{ 'remote-container': remoteMuted }">
+      <button
+        *ngIf="remoteMuted"
+        id="remote-muted-button"
+        data-module="govuk-button"
+        type="button"
+        class="govuk-button govuk-button--warning reduce-margin hand"
+        aria-label="Remote muted button"
+      >
+        The judge muted you
+        <span class="govuk-button__start-icon microphoneIcon">
+          <img id="unmute-microphone-img" class="microphoneIcon" src="/assets/images/mic_off.png" alt="Unmute Microphone Image" />
+        </span>
+      </button>
+
+      <!-- toggle hand raise/lower-->
+
+      <button
+        *ngIf="!isPrivateConsultation"
+        id="toggle-hand-button"
+        data-module="govuk-button"
+        type="button"
+        class="govuk-button reduce-margin hand"
+        [ngClass]="handRaised ? 'govuk-button--warning' : ''"
+        aria-label="Toggle hand raised status"
+        (click)="toggleHandRaised()"
+      >
+        {{ handToggleText }}
+        <span class="govuk-button__start-icon">
+          <img
+            id="toggle-hand-img"
+            class="microphoneIcon"
+            [src]="handRaised ? '/assets/images/hand.png' : '/assets/images/no-hand.png'"
+            alt="Raise"
+          />
+        </span>
+      </button>
+    </span>
+  </div>
+</div>
