+++ VideoWeb/VideoWeb/ClientApp/src/app/waiting-space/hearing-controls/hearing-controls.component.scss
@@ -1,0 +1,58 @@
+#outgoingFeedDiv {
+  left: 0.2%;
+  position: fixed;
+  bottom: 1%;
+}
+
+#outgoingFeedVideo {
+  left: 1%;
+  position: fixed;
+  bottom: 5.5vh;
+  max-width: 13%;
+  min-width: 15%;
+  transform: rotateY(180deg);
+  -webkit-transform: rotateY(180deg); /* Safari and Chrome */
+  -moz-transform: rotateY(180deg); /* Firefox */
+}
+
+.hideVideo {
+  opacity: 0;
+}
+
+.showVideo {
+  opacity: 1;
+}
+
+.microphoneIcon {
+  height: 18px;
+}
+
+.reduce-margin {
+  margin-bottom: 2px;
+}
+
+#muteButton .govuk-button:focus {
+  background-color: govuk-colour('green');
+}
+
+#unmuteButton .govuk-button--warning:focus {
+  background-color: govuk-colour('red');
+}
+
+.hand {
+  margin-left: 1vw;
+}
+
+#toggle-hand-button.govuk-button:focus {
+  background-color: govuk-colour('green');
+  color: govuk-colour('white');
+}
+
+#toggle-hand-button.govuk-button--warning:focus {
+  background-color: govuk-colour('red');
+  color: govuk-colour('white');
+}
+
+.remote-container {
+  margin-left: -1vw;
+}
