+++ VideoWeb/VideoWeb/ClientApp/src/app/waiting-space/hearing-controls/hearing-controls.component.spec.ts
@@ -1,0 +1,174 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { HearingControlsComponent } from './hearing-controls.component';
+import { ConferenceTestData } from 'src/app/testing/mocks/data/conference-test-data';
+import { ConferenceResponse, Role, ParticipantStatus } from 'src/app/services/clients/api-client';
+import { eventsServiceSpy, participantStatusSubjectMock } from 'src/app/testing/mocks/mock-events-service';
+import { videoCallServiceSpy, onParticipantUpdatedMock } from 'src/app/testing/mocks/mock-video-call-service';
+import { Logger } from 'src/app/services/logging/logger-base';
+import { MockLogger } from 'src/app/testing/mocks/MockLogger';
+import { ParticipantUpdated } from '../models/video-call-models';
+import { ParticipantStatusMessage } from 'src/app/services/models/participant-status-message';
+
+describe('HearingControlsComponent', () => {
+    let component: HearingControlsComponent;
+    const gloalConference = new ConferenceTestData().getConferenceDetailPast() as ConferenceResponse;
+    const globalParticipant = gloalConference.participants.filter(x => x.role === Role.Individual)[0];
+
+    const eventsService = eventsServiceSpy;
+    const participantStatusSubject = participantStatusSubjectMock;
+
+    const videoCallService = videoCallServiceSpy;
+    const onParticipantUpdatedSubject = onParticipantUpdatedMock;
+
+    const logger: Logger = new MockLogger();
+
+    beforeEach(() => {
+        component = new HearingControlsComponent(videoCallService, eventsService, logger);
+        component.participant = globalParticipant;
+        component.ngOnInit();
+    });
+
+    afterEach(() => {
+        component.ngOnDestroy();
+    });
+
+    it('should raise hand on toggle if hand not raised', () => {
+        component.handRaised = false;
+        component.toggleHandRaised();
+        expect(videoCallService.raiseHand).toHaveBeenCalledTimes(1);
+        expect(component.handToggleText).toBe('Lower my hand');
+    });
+
+    it('should lower hand on toggle if hand raised', () => {
+        component.handRaised = true;
+        component.toggleHandRaised();
+        expect(videoCallService.lowerHand).toHaveBeenCalledTimes(1);
+        expect(component.handToggleText).toBe('Raise my hand');
+    });
+
+    it('should show raised hand on hand lowered', () => {
+        const payload = new ParticipantUpdated('YES', 0, globalParticipant.tiled_display_name);
+        onParticipantUpdatedSubject.next(payload);
+        expect(component.remoteMuted).toBeTruthy();
+        expect(component.handRaised).toBeFalsy();
+        expect(component.handToggleText).toBe('Raise my hand');
+    });
+
+    it('should not show raised hand on hand lowered for another participant', () => {
+        const otherParticipant = gloalConference.participants.filter(x => x.role === Role.Representative)[0];
+        const payload = new ParticipantUpdated('YES', 0, otherParticipant.tiled_display_name);
+        component.handRaised = true;
+        component.remoteMuted = false;
+        onParticipantUpdatedSubject.next(payload);
+        expect(component.remoteMuted).toBeFalsy();
+        expect(component.handRaised).toBeTruthy();
+        expect(component.handToggleText).toBe('Lower my hand');
+    });
+
+    it('should show lower hand on hand raised', () => {
+        const payload = new ParticipantUpdated('YES', 123, globalParticipant.tiled_display_name);
+        onParticipantUpdatedSubject.next(payload);
+        expect(component.handRaised).toBeTruthy();
+        expect(component.handToggleText).toBe('Lower my hand');
+    });
+
+    it('should not show lower hand when hand raised for another participant', () => {
+        const otherParticipant = gloalConference.participants.filter(x => x.role === Role.Representative)[0];
+        const payload = new ParticipantUpdated('YES', 0, otherParticipant.tiled_display_name);
+        component.handRaised = false;
+        onParticipantUpdatedSubject.next(payload);
+        expect(component.handRaised).toBeFalsy();
+        expect(component.handToggleText).toBe('Raise my hand');
+    });
+
+    it('should mute locally if remote muted and not muted locally', () => {
+        videoCallService.toggleMute.calls.reset();
+        const payload = new ParticipantUpdated('YES', 123, globalParticipant.tiled_display_name);
+        component.audioMuted = false;
+        component.handleParticipantUpdatedInVideoCall(payload);
+        expect(videoCallService.toggleMute).toHaveBeenCalledTimes(1);
+    });
+
+    it('should skip mute locally if remote muted and already muted locally', () => {
+        videoCallService.toggleMute.calls.reset();
+        const payload = new ParticipantUpdated('YES', 123, globalParticipant.tiled_display_name);
+        component.audioMuted = true;
+        component.handleParticipantUpdatedInVideoCall(payload);
+        expect(videoCallService.toggleMute).toHaveBeenCalledTimes(0);
+    });
+
+    it('should not reset mute when participant status to available', () => {
+        spyOn(component, 'resetMute').and.callThrough();
+        const status = ParticipantStatus.Available;
+        const message = new ParticipantStatusMessage(globalParticipant.id, globalParticipant.username, gloalConference.id, status);
+
+        participantStatusSubject.next(message);
+
+        expect(component.resetMute).toHaveBeenCalledTimes(0);
+    });
+
+    it('should reset mute when participant status to in consultation', () => {
+        spyOn(component, 'resetMute').and.callThrough();
+        const status = ParticipantStatus.InConsultation;
+        const participant = globalParticipant;
+        const message = new ParticipantStatusMessage(participant.id, participant.username, gloalConference.id, status);
+
+        participantStatusSubject.next(message);
+
+        expect(component.resetMute).toHaveBeenCalled();
+    });
+
+    it('should ignore participant updates for another participant', () => {
+        spyOn(component, 'resetMute').and.callThrough();
+        const status = ParticipantStatus.InConsultation;
+        const participant = gloalConference.participants.filter(x => x.role === Role.Representative)[0];
+        const message = new ParticipantStatusMessage(participant.id, participant.username, gloalConference.id, status);
+
+        participantStatusSubject.next(message);
+
+        expect(component.resetMute).toHaveBeenCalledTimes(0);
+    });
+
+    it('should show self view on-click when currently hidden', () => {
+        component.selfViewOpen = false;
+        component.toggleView();
+        expect(component.selfViewOpen).toBeTruthy();
+    });
+
+    it('should hide self view on-click when currently visible', () => {
+        component.selfViewOpen = true;
+        component.toggleView();
+        expect(component.selfViewOpen).toBeFalsy();
+    });
+
+    it('should mute the participant when user opts to mute the call', () => {
+        videoCallService.toggleMute.and.returnValue(true);
+        component.toggleMute();
+        expect(component.audioMuted).toBeTruthy();
+    });
+
+    it('should unmute the participant when user opts to turn off mute option', () => {
+        videoCallService.toggleMute.and.returnValue(false);
+        component.toggleMute();
+        expect(component.audioMuted).toBeFalsy();
+    });
+
+    it('should unmute the participant already muted', () => {
+        spyOn(component, 'toggleMute').and.callThrough();
+        videoCallService.toggleMute.and.returnValue(false);
+        component.audioMuted = true;
+        component.resetMute();
+        expect(videoCallService.toggleMute).toHaveBeenCalled();
+        expect(component.toggleMute).toHaveBeenCalled();
+        expect(component.audioMuted).toBeFalsy();
+    });
+
+    it('should not reset mute option the participant not in mute', () => {
+        spyOn(component, 'toggleMute').and.callThrough();
+        component.audioMuted = false;
+        component.resetMute();
+        expect(component.toggleMute).toHaveBeenCalledTimes(0);
+        expect(component.audioMuted).toBeFalsy();
+    });
+});
