+++ VideoWeb/VideoWeb/ClientApp/src/app/waiting-space/participant-network-health/participant-network-health.component.ts
@@ -17,7 +17,6 @@ export class ParticipantNetworkHealthComponent extends ParticipantNetworkHealthB
 
     constructor(protected eventsService: EventsService, private modalService: ModalService) {
         super(eventsService);
-        console.log('waiting room monitor loaded');
     }
 
     get isNetworkGood() {
