+++ VideoWeb/VideoWeb/ClientApp/src/app/waiting-space/participant-waiting-room/participant-waiting-room.component.scss
@@ -67,37 +67,8 @@
   margin-top: -20px;
 }
 
-#outgoingFeedDiv {
-  left: 0.2%;
-  position: fixed;
-  bottom: 1%;
-}
-
-#outgoingFeedVideo {
-  left: 1%;
-  position: fixed;
-  bottom: 5.5vh;
-  max-width: 13%;
-  min-width: 15%;
-  transform: rotateY(180deg);
-  -webkit-transform: rotateY(180deg); /* Safari and Chrome */
-  -moz-transform: rotateY(180deg); /* Firefox */
-}
-
 .reduce-margin {
   margin-bottom: 2px;
-}
-
-.hideVideo {
-  opacity: 0;
-}
-
-.showVideo {
-  opacity: 1;
-}
-
-.microphoneIcon {
-  height: 18px;
 }
 
 .no-mass {
